requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.98.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.42
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
